Move Duke.java into package duke.main
